feat: add goalInput component (33)
