+++ rn-complete-guide/App.js
@@ -10,33 +10,21 @@ import {
 } from "react-native";
 
 import GoalItem from "./components/GoalItem";
+import GoalInput from "./components/GoalInput";
 
 export default function App() {
-  const [enteredGoal, setEnteredGoal] = useState(""); // пустая строка, т.к. юзер еще ничего не ввел
   const [courseGoals, setCourseGoals] = useState([]);
 
-  const goalInputHandler = (enteredText) => {
-    setEnteredGoal(enteredText);
-  };
-
-  const addGoalHandler = () => {
+  const addGoalHandler = (goalTitle) => {
     setCourseGoals((currentGoals) => [
       ...currentGoals,
-      { id: Math.random().toString(), value: enteredGoal },
+      { id: Math.random().toString(), value: goalTitle },
     ]); 
   };
 
   return (
     <View style={styles.root}>
-      <View style={styles.inputContainer}>
-        <TextInput
-          placeholder="Course Goal"
-          onChangeText={goalInputHandler}
-          value={enteredGoal}
-          style={styles.input}
-        />
-        <Button title="ADD" onPress={addGoalHandler} />
-      </View>
+      <GoalInput onAddGoal={addGoalHandler} />
       <FlatList
         keyExtractor={(item, index) => item.id}
         data={courseGoals}
@@ -54,12 +42,5 @@ const styles = StyleSheet.create({
     flexDirection: "row",
     justifyContent: "space-between",
     alignItems: "center",
-  },
-  input: {
-    width: "80%",
-    borderColor: "black",
-    borderBottomWidth: 1,
-    padding: 10,
-    marginBottom: 5,
   },
 });
