+++ rn-complete-guide/components/GoalInput.js
@@ -1,22 +1,47 @@
-import React from "react";
-import { Text, View, StyleSheet } from "react-native";
+import React, {useState} from "react";
+import {
+  StyleSheet,
+  Text,
+  View,
+  TextInput,
+  Button,
+  ScrollView,
+  FlatList,
+} from "react-native";
 
-const GoalItem = (props) => {
+const GoalInput = (props) => {
+  const [enteredGoal, setEnteredGoal] = useState(""); 
+
+  const goalInputHandler = (enteredText) => {
+    setEnteredGoal(enteredText);
+  };
+
   return (
-    <View style={styles.listItem}>
-      <Text>{props.title}</Text>
-    </View>
+  <View style={styles.inputContainer}>
+    <TextInput
+      placeholder="Course Goal"
+      onChangeText={goalInputHandler}
+      value={enteredGoal}
+      style={styles.input}
+    />
+    <Button title="ADD" onPress={props.onAddGoal.bind(this, enteredGoal)} />
+  </View>
   );
 };
 
 const styles = StyleSheet.create({
-  listItem: {
-    padding: 10,
-    marginVertical: 10,
-    backgroundColor: "#CCC",
+  inputContainer: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  input: {
+    width: "80%",
     borderColor: "black",
-    borderWidth: 1,
+    borderBottomWidth: 1,
+    padding: 10,
+    marginBottom: 5,
   },
 });
 
-export default GoalItem;
+export default GoalInput;
